demo3/requirements.txt
beautifulsoup4==4.11.1
google-cloud-bigquery==2.34.3
google-cloud-bigquery-storage==2.13.1
oogle-cloud-language==2.4.2
multiprocess==0.70.13
pandarallel==1.6.1
pandas==1.3.5
ratelimit==2.2.1
requests==2.27.1
seaborn==0.11.2
tqdm==4.64.0